PLF-8190 : add interface LoginHistoryDAO and classes LoginHistoryStorage JPALoginHistoryStorageImpl JCRLoginHistoryStorageImpl and LoginHistoryEntity constructor and empty LoginHistoryServiceImpl methods
