+++ exo-gadget-pack/gadget-pack-services/src/main/java/org/exoplatform/platform/gadget/services/LoginHistory/jpa/dao/LoginHistoryDAO.java
@@ -1,0 +1,6 @@
+package org.exoplatform.platform.gadget.services.LoginHistory.jpa.dao;
+
+import org.exoplatform.commons.persistence.impl.GenericDAOJPAImpl;
+
+public class LoginHistoryDAO extends GenericDAOJPAImpl {
+}
